Correct package declarations
