+++ datacapture/src/main/java/com/google/android/fhir/datacapture/MoreContexts.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.utilities
+package com.google.android.fhir.datacapture
 
 import android.content.Context
 import androidx.appcompat.app.AppCompatActivity
