+++ datacapture/src/main/java/com/google/android/fhir/datacapture/MoreLocalDateTimes.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.utilities
+package com.google.android.fhir.datacapture
 
 import android.content.Context
 import android.text.format.DateFormat
