+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreContextsTest.kt
@@ -19,7 +19,6 @@ package com.google.android.fhir.datacapture
 import android.app.Application
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.view.ContextThemeWrapper
-import com.google.android.fhir.datacapture.utilities.tryUnwrapContext
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
