+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreEnumerationTest.kt
@@ -16,7 +16,6 @@
 
 package com.google.android.fhir.datacapture
 
-import com.google.android.fhir.datacapture.utilities.toCoding
 import com.google.common.truth.Truth.assertThat
 import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.Enumeration
