+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreLocalDateTimesTest.kt
@@ -18,8 +18,6 @@ package com.google.android.fhir.datacapture
 
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
-import com.google.android.fhir.datacapture.utilities.format
-import com.google.android.fhir.datacapture.utilities.toLocalizedTimeString
 import com.google.common.truth.Truth.assertThat
 import java.time.LocalDate
 import java.time.LocalDateTime
