+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreLocalTimesTest.kt
@@ -18,7 +18,6 @@ package com.google.android.fhir.datacapture
 
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
-import com.google.android.fhir.datacapture.utilities.toLocalizedString
 import com.google.common.truth.Truth.assertThat
 import java.time.LocalTime
 import java.util.Locale
