+++ datacapture/src/test/java/com/google/android/fhir/datacapture/MoreReflectionUtilTest.kt
@@ -16,7 +16,6 @@
 
 package com.google.android.fhir.datacapture
 
-import com.google.android.fhir.datacapture.utilities.invokeFunction
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
