+++ datacapture/src/test/java/com/google/android/fhir/datacapture/TypeConversionUtilTest.kt
@@ -16,9 +16,6 @@
 
 package com.google.android.fhir.datacapture
 
-import com.google.android.fhir.datacapture.utilities.toCodeType
-import com.google.android.fhir.datacapture.utilities.toIdType
-import com.google.android.fhir.datacapture.utilities.toUriType
 import com.google.common.truth.Truth.assertThat
 import org.hl7.fhir.r4.model.CodeType
 import org.hl7.fhir.r4.model.Coding
